requirements.txt
boto3==1.39.4
botocore==1.39.4
email-validator==2.2.0
fastapi==0.116.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.8.2
sqlalchemy==2.0.41
starlette==0.46.2
typing_extensions==4.14.1
uvicorn==0.35.0
databases[asyncpg]==0.9.0
psycopg2-binary==2.9.10
python-multipart==0.0.20
alembic==1.16.3
httpx==0.28.1
gunicorn==23.0.0
sentry-sdk[fastapi]==2.32.0
phonenumbers==9.0.9
greenlet==3.2.3
jmespath==1.0.1
packaging==25.0
passlib==1.7.4
python-jose==3.5.0
rsa==4.9.1
s3transfer==0.13.0
validators==0.35.0
redis[hiredis]==5.2.0

# AI and ML Libraries
google-genai==1.25.0
groq==0.29.0


cryptography==46.0.3